requirements.txt
alembic==1.7.4
bcrypt==3.2.0
cffi==1.14.6
click==6.6
colorama==0.4.4
coverage==4.2
cryptography==35.0.0
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==2.0.2
Flask-MySQLdb==0.2.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.3
protobuf==3.18.1
pycparser==2.17
PyJWT==1.4.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.3
pytz==2021.3
six==1.10.0
SQLAlchemy==1.3.24
Werkzeug==0.16.1
zipp==3.6.0
gunicorn==20.1.0
requests==2.23.0
websocket-client==0.57.0
websockets==8.1
python-dotenv==0.17.0